Update: move component directory - server => client
